pyproject.toml
[tool.poetry]
name = "sleepy-bot"
version = "0.1.0"
description = ""
authors = ["pavel"]
package-mode = false


[tool.poetry.dependencies]
# Основные зависимости
python = "^3.13"

# Веб/Боты (Aiogram)
aiogram = "^3.20.0.post0"
aiogram-dialog = "^2.3.1"

# Валидация и конфиги (Pydantic)
pydantic = {extras = ["email"], version = "^2.11.5"}
pydantic-settings = "^2.9.1"

# База данных (PostgreSQL + ORM)
asyncpg = "^0.30.0"
tortoise-orm = "^0.25.0"
aerich = "^0.9.0"

# Логирование
loguru = "^0.7.3"

# DI (Dependency Injection)
dishka = "^1.6.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^6.0.1"
pre-commit = "^4.2.0"
pyperclip = "^1.9.0"
gitprompter = "^0.1.0"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
mypy = "^1.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 95
skip-string-normalization = true
preview = true


[tool.isort]
profile = "black"
line_length = 95
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
#disable_error_code = import-not-found
warn_return_any = true # функции должны возвращать конкретные типы
disallow_untyped_defs = true # сигнатура функций должна быть типизирована
ignore_missing_imports = true # ингорировать жалобы в роде: не хватает библиотеки types-{lib_name}
explicit_package_bases = true
exclude = [
    "tests/",
    "infra/db/migrate_script.py",
]
